requirements.txt
cryptography>=42.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
psutil>=5.0.0
